% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{create_database}
\alias{create_database}
\title{create a database}
\usage{
create_database(con, DBname)
}
\arguments{
\item{con}{the RMySQL connection}

\item{DBname}{the RMySQL Database Name}
}
\description{
create a database
}
