% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{create_table}
\alias{create_table}
\title{Create table}
\usage{
create_table(con, tab, colnams, coltypes, keycols, add_keycol = NULL)
}
\arguments{
\item{con}{the RMySQL connection}

\item{tab}{the table name string}

\item{colnams}{the names of the columns}

\item{coltypes}{the types of the columns}

\item{keycols}{the colnums that will form indexes}
}
\description{
Create table
}
