% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{create_table_df}
\alias{create_table_df}
\title{Create table based on a dataframe}
\usage{
create_table_df(con, tab, df, keycols, unstr = TRUE)
}
\arguments{
\item{con}{the RMySQL connection}

\item{tab}{the table name string}

\item{df}{the dataframe}

\item{keycols}{the colnums that will form indexes}

\item{unstr}{determine the structure to be unstructured}
}
\description{
Create table based on a dataframe
}
