% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{droptables}
\alias{droptables}
\title{drop tables on the connection}
\usage{
droptables(con, tab)
}
\arguments{
\item{con}{the RMySQL connection}

\item{tab}{the table name string}
}
\description{
drop tables on the connection
}
