% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL upload.R
\name{get_table}
\alias{get_table}
\title{get tables as data frame}
\usage{
get_table(pool, tab)
}
\arguments{
\item{tab}{the table name string}

\item{con}{the RMySQL connection}
}
\description{
get tables as data frame
}
