% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{insert_table_df}
\alias{insert_table_df}
\title{Insert data onto SQL server}
\usage{
insert_table_df(con, tab, df, file_loc = NULL)
}
\arguments{
\item{con}{the RMySQL connection}

\item{tab}{the table name string}

\item{df}{the dataframe}

\item{file_loc}{the location to write the file to harddrive, when NULL it would be a temporary location on local drive}
}
\description{
Insert data onto SQL server
}
