% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL upload.R
\name{insert_value}
\alias{insert_value}
\title{insert values based on data frame}
\usage{
insert_value(pool, tab, df)
}
\arguments{
\item{tab}{the table name string}

\item{df}{the data frame to update values}

\item{con}{the RMySQL connection}
}
\description{
insert values based on data frame
}
