% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{push_table_df}
\alias{push_table_df}
\title{push a data frame onto SQL server, create new table if doesn't exist}
\usage{
push_table_df(con, tbl, df, keycols, ...)
}
\arguments{
\item{con}{the RMySQL connection}

\item{df}{the dataframe}

\item{tab}{the table name string}
}
\description{
push a data frame onto SQL server, create new table if doesn't exist
}
