% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL Maintainence.R
\name{show_indexes}
\alias{show_indexes}
\title{show indexes of a table}
\usage{
show_indexes(con, tab)
}
\arguments{
\item{con}{the RMySQL connection}

\item{tab}{the table name string}
}
\description{
show indexes of a table
}
