% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SQL upload.R
\name{UPDATE_value}
\alias{UPDATE_value}
\title{update values based on data frame}
\usage{
UPDATE_value(con, tab, df, val_fields, pk_fields)
}
\arguments{
\item{con}{the RMySQL connection}

\item{tab}{the table name string}

\item{df}{the data frame to update values}

\item{val_fields}{the value fields to be updated}

\item{pk_fields}{the fields that locate unique rows}
}
\description{
update values based on data frame
}
